+++ pyesg/__init__.py
@@ -1,6 +1,6 @@
 """Top level imports"""
-from pyesg.interpolator import NelsonSiegelSvensson
 from pyesg.interpolators.nelson_siegel import NelsonSiegel
+from pyesg.interpolators.nelson_siegel_svensson import NelsonSiegelSvensson
 
 from pyesg.processes.academy_rate_process import AcademyRateProcess
 from pyesg.processes.black_scholes_process import BlackScholesProcess
